1.19.9 config improvement chorus. package move
