+++ src/main/java/com/lothrazar/cyclicmagic/module/MultiContent.java
@@ -81,8 +81,6 @@ import com.lothrazar.cyclicmagic.guide.GuideCategory;
 import com.lothrazar.cyclicmagic.guide.GuideItem;
 import com.lothrazar.cyclicmagic.guide.GuideRegistry;
 import com.lothrazar.cyclicmagic.item.ItemHeartContainer;
-import com.lothrazar.cyclicmagic.item.buildswap.ItemBuildSwapper;
-import com.lothrazar.cyclicmagic.item.buildswap.ItemBuildSwapper.WandType;
 import com.lothrazar.cyclicmagic.item.core.BaseItemProjectile;
 import com.lothrazar.cyclicmagic.item.dynamite.EntityDynamiteBlockSafe;
 import com.lothrazar.cyclicmagic.item.dynamite.EntityDynamiteMining;
@@ -112,6 +110,8 @@ import com.lothrazar.cyclicmagic.item.equipmentsandstone.ItemSandstoneHoe;
 import com.lothrazar.cyclicmagic.item.equipmentsandstone.ItemSandstonePickaxe;
 import com.lothrazar.cyclicmagic.item.equipmentsandstone.ItemSandstoneSpade;
 import com.lothrazar.cyclicmagic.item.equipmentsandstone.ItemSandstoneSword;
+import com.lothrazar.cyclicmagic.item.exchangebuild.ItemBuildSwapper;
+import com.lothrazar.cyclicmagic.item.exchangebuild.ItemBuildSwapper.WandType;
 import com.lothrazar.cyclicmagic.item.firemagic.EntityBlazeBolt;
 import com.lothrazar.cyclicmagic.item.firemagic.ItemProjectileBlaze;
 import com.lothrazar.cyclicmagic.item.minecart.EntityGoldFurnaceMinecart;
