+++ src/main/java/com/lothrazar/cyclicmagic/registry/PacketRegistry.java
@@ -28,7 +28,6 @@ import com.lothrazar.cyclicmagic.block.exppylon.PacketTilePylon;
 import com.lothrazar.cyclicmagic.block.password.PacketTilePassword;
 import com.lothrazar.cyclicmagic.block.vector.PacketTileVector;
 import com.lothrazar.cyclicmagic.gui.container.PacketGuiShortOverride;
-import com.lothrazar.cyclicmagic.item.buildswap.PacketSwapBlock;
 import com.lothrazar.cyclicmagic.item.cannon.MessageGolemLaserFX;
 import com.lothrazar.cyclicmagic.item.cyclicwand.PacketSpellBuildSize;
 import com.lothrazar.cyclicmagic.item.cyclicwand.PacketSpellFromServer;
@@ -38,6 +37,7 @@ import com.lothrazar.cyclicmagic.item.cyclicwand.PacketWandGui;
 import com.lothrazar.cyclicmagic.item.enderbook.PacketDeleteWaypoint;
 import com.lothrazar.cyclicmagic.item.enderbook.PacketNewButton;
 import com.lothrazar.cyclicmagic.item.enderbook.PacketWarpButton;
+import com.lothrazar.cyclicmagic.item.exchangebuild.PacketSwapBlock;
 import com.lothrazar.cyclicmagic.item.merchant.PacketSyncVillagerToClient;
 import com.lothrazar.cyclicmagic.item.merchant.PacketSyncVillagerToServer;
 import com.lothrazar.cyclicmagic.item.merchant.PacketVillagerTrade;
